requirements.txt
dune-client==1.3.0
pyyaml
python-dotenv
pandas